feat!: Packaging changes.
Specifically:

* Remove UMD module support. Per caniuse.com, all browsers that
support WebAssembly also support ES6, save Firefox < 54, which is now
over 5 years old.

* Add `files` to the package config and ensure that all files in
`dist` and `wasm_dist` are included in the package.

Note that Storybook expects CommonJS format, and adding `type: module`
to `package.json` causes it to break. This will be fixed in Storybook
v7, but until then, we implement the workaround described here:

https://github.com/storybookjs/storybook/issues/11587#issuecomment-1243797074