refactor: Multiple tsconfigs.
Move common settings into a base config, then extend it as follows:

* Whole-project typecheck-only config which emits nothing.

* Typecheck-only config for just the library (no Storybook, etc.),
which emits only type declarations.

Additionally, we make the following changes to all configs:

* Remove `downlevelIteration` setting as we're not using TypeScript to
emit code, only types.

* Remove `lib` setting, as it's redundant: "DOM" is included by
default, and the appropriate module types are included according to
the `target` setting.

* Target ES6, which should work on all browsers which have Wasm
support, and is sufficient for this module's needs.

* Remove ES2018 targets, which are unnecessary for this module.