chore: Add vite-plugin-checker.
This enables Vite to run the TypeScript build concurrently with the
Vite build, which speeds up the process.

Note that due to a bug in vite-plugin-checker that prevents setting
`tsconfigPath` when `buildMode` is true, we need to rename
`tsconfig.type.json` to `tsconfig.json`, as this is the config we want
to use when building the module. The old `tsconfig.json` config is
renamed `tsconfig.all.json` and is now used explicitly in the
`typecheck` script.

https://github.com/fi3ework/vite-plugin-checker/issues/181